feat(DatePicker): Replace deprecated implementation
