+++ packages/react-component-library/src/components/DatePicker/DatePicker.test.tsx
@@ -1,36 +1,40 @@
+import { format, isValid } from 'date-fns'
 import React, { useState } from 'react'
 import '@testing-library/jest-dom/extend-expect'
-import { ColorDanger600, ColorSuccess600 } from '@defencedigital/design-tokens'
+import { ColorDanger800 } from '@defencedigital/design-tokens'
 import {
   fireEvent,
   render,
   RenderResult,
   waitFor,
+  waitForElementToBeRemoved,
 } from '@testing-library/react'
 import 'jest-styled-components'
 import userEvent from '@testing-library/user-event'
 
-import { DatePicker } from '.'
-import { Button } from '../Button'
+import { BORDER_WIDTH } from '../../styled-components'
+import { COMPONENT_SIZE } from '../Forms'
+import { DatePicker, DatePickerOnChangeData } from '.'
+import { DATE_VALIDITY } from './constants'
 
 const NOW = '2019-12-05T11:00:00.000Z'
-const ERROR_BORDER = `1px solid ${ColorDanger600}`
+const ERROR_BOX_SHADOW = `0 0 0 ${
+  BORDER_WIDTH[COMPONENT_SIZE.FORMS]
+} ${ColorDanger800.toUpperCase()}`
 
-function click(element: HTMLElement) {
-  fireEvent.mouseDown(element)
-  fireEvent.click(element)
+function formatDate(date: Date | null) {
+  return date && isValid(date) ? format(date, 'dd/MM/yyyy') : ''
 }
 
 describe('DatePicker', () => {
   let wrapper: RenderResult
-  let startDate: Date
+  let initialStartDate: Date
   let label: string
-  let onChange: (data: { startDate: Date; endDate: Date }) => void
   let onBlur: (e: React.FormEvent) => void
   let onCalendarFocus: (e: React.SyntheticEvent) => void
-  let dateSpy: jest.SpyInstance
   let days: string[]
   let onSubmitSpy: (e: React.FormEvent) => void
+  const onChange = jest.fn<void, [DatePickerOnChangeData]>()
 
   function assertInputButtonAria({
     ariaExpanded,
@@ -48,10 +52,6 @@ describe('DatePicker', () => {
   }
 
   beforeAll(() => {
-    dateSpy = jest
-      .spyOn(Date, 'now')
-      .mockImplementation(() => new Date(NOW).valueOf())
-
     function leadingZero(n: number): string {
       return n > 9 ? `${n}` : `0${n}`
     }
@@ -59,17 +59,26 @@ describe('DatePicker', () => {
     days = new Array(31).map((i) => leadingZero(i + 1)) // [01, 02, ..., 31]
   })
 
+  beforeEach(() => {
+    jest.useFakeTimers()
+    jest.setSystemTime(Date.parse(NOW))
+  })
+
+  afterEach(() => {
+    jest.useRealTimers()
+    onChange.mockReset()
+  })
+
   describe('default props', () => {
     beforeEach(() => {
-      startDate = new Date(2019, 11, 1)
-      onChange = jest.fn()
+      initialStartDate = new Date(2019, 11, 1)
       onBlur = jest.fn()
       onCalendarFocus = jest.fn()
 
       wrapper = render(
         <>
           <DatePicker
-            startDate={startDate}
+            initialStartDate={initialStartDate}
             onChange={onChange}
             onBlur={onBlur}
             onCalendarFocus={onCalendarFocus}
@@ -158,7 +167,21 @@ describe('DatePicker', () => {
           return waitFor(() => {
             expect(
               wrapper.queryByTestId('floating-box-content')
-            ).not.toBeVisible()
+            ).not.toBeInTheDocument()
+          })
+        })
+      })
+
+      describe('and the input is clicked on', () => {
+        beforeEach(() => {
+          userEvent.click(wrapper.getByTestId('datepicker-input'))
+        })
+
+        it('hides the day picker container', () => {
+          return waitFor(() => {
+            expect(
+              wrapper.queryByTestId('floating-box-content')
+            ).not.toBeInTheDocument()
           })
         })
       })
@@ -169,65 +192,24 @@ describe('DatePicker', () => {
         wrapper.getByTestId('datepicker-input').focus()
       })
 
-      it('shows the month', () => {
-        expect(wrapper.queryByText('December 2019')).toBeInTheDocument()
-      })
-
-      it('displays the container', () => {
+      it("doesn't display the day picker", () => {
         return waitFor(() => {
-          expect(wrapper.getByTestId('floating-box')).toBeVisible()
+          expect(wrapper.queryByTestId('floating-box')).not.toBeInTheDocument()
         })
       })
+    })
 
-      describe('and clicks on a day', () => {
-        beforeEach(() => {
-          click(wrapper.getByText('31'))
-        })
-
-        it('set the value of the component to this date', () => {
-          expect(
-            wrapper.getByTestId('datepicker-input').getAttribute('value')
-          ).toBe('31/12/2019')
-        })
-
-        it('invokes the onChange callback', () => {
-          expect(onChange).toHaveBeenCalledTimes(1)
-          expect(onChange).toHaveBeenCalledWith({
-            startDate: new Date('2019-12-31T12:00:00.000Z'),
-            endDate: new Date('2019-12-31T12:00:00.000Z'),
-          })
-        })
-
-        it('hides the container', () => {
-          return waitFor(() => {
-            expect(wrapper.getByTestId('floating-box')).not.toBeVisible()
-          })
-        })
+    describe('when the input is cleared using the keyboard', () => {
+      beforeEach(() => {
+        userEvent.clear(wrapper.getByTestId('datepicker-input'))
       })
 
-      describe('and clicks outside the date picker', () => {
-        beforeEach(() => {
-          userEvent.click(wrapper.getByTestId('datepicker-outside'))
-        })
-
-        it('hides the day picker container', () => {
-          return waitFor(() => {
-            expect(
-              wrapper.queryByTestId('floating-box-content')
-            ).not.toBeVisible()
-          })
-        })
-
-        describe('and the user focuses on the input', () => {
-          beforeEach(() => {
-            userEvent.click(wrapper.getByTestId('datepicker-input-button'))
-          })
-
-          it('shows the day picker container', () => {
-            return waitFor(() => {
-              expect(wrapper.getByTestId('floating-box')).toBeVisible()
-            })
-          })
+      it('calls the `onChange` callback with null dates', () => {
+        expect(onChange).toHaveBeenLastCalledWith({
+          startDate: null,
+          startDateValidity: null,
+          endDate: null,
+          endDateValidity: null,
         })
       })
     })
@@ -235,42 +217,43 @@ describe('DatePicker', () => {
     describe('when the end user types a new date value', () => {
       describe('and the date value is valid', () => {
         describe('and the day/month use two digit format', () => {
-          beforeEach(async () => {
+          beforeEach(() => {
             const input = wrapper.getByTestId('datepicker-input')
 
-            await userEvent.type(input, '{backspace}01/05/2016')
+            userEvent.type(input, '{selectall}01/05/2016')
           })
 
           it('set the value of the component to this date', () => {
-            expect(
-              wrapper.getByTestId('datepicker-input').getAttribute('value')
-            ).toBe('01/05/2016')
+            expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+              '01/05/2016'
+            )
           })
 
           it('invokes the `onChange` callback', () => {
-            expect(onChange).toHaveBeenCalledTimes(1)
-            expect(onChange).toHaveBeenCalledWith({
+            expect(onChange).toHaveBeenLastCalledWith({
               startDate: new Date('2016-05-01T12:00:00.000Z'),
+              startDateValidity: DATE_VALIDITY.VALID,
               endDate: new Date('2016-05-01T12:00:00.000Z'),
+              endDateValidity: DATE_VALIDITY.VALID,
             })
           })
 
-          it('updates the selected day', () => {
-            expect(wrapper.getByText('May 2016')).toBeInTheDocument()
-            expect(wrapper.getByText('1')).toBeInTheDocument()
-          })
+          describe('and the day picker is opened', () => {
+            beforeEach(() => {
+              userEvent.click(wrapper.getByTestId('datepicker-input-button'))
+            })
 
-          it('hides the day picker container', () => {
-            return waitFor(() => {
-              expect(wrapper.getByTestId('floating-box')).not.toBeVisible()
+            it('displays the new date', () => {
+              expect(wrapper.getByText('May 2016')).toBeInTheDocument()
+              expect(wrapper.getByText('1')).toBeInTheDocument()
             })
           })
 
           describe('and the tab key is pressed', () => {
-            beforeEach(async () => {
+            beforeEach(() => {
               wrapper.getByTestId('datepicker-input').focus()
 
-              await userEvent.tab()
+              userEvent.tab()
             })
 
             it('focuses the picker open/close button', () => {
@@ -280,156 +263,152 @@ describe('DatePicker', () => {
             })
 
             describe('and the space key is pressed', () => {
-              beforeEach(async () => {
+              beforeEach(() => {
                 const button = wrapper.getByTestId('datepicker-input-button')
 
-                await userEvent.type(button, '{space}', { skipClick: true })
+                userEvent.type(button, '{space}', { skipClick: true })
               })
 
-              it('closes the picker container', () => {
+              it('opens the picker container', () => {
                 expect(
                   wrapper.getByTestId('datepicker-input-button')
-                ).toHaveAttribute('aria-label', 'Show day picker')
-              })
-            })
-
-            describe('and the tab  key is pressed again', () => {
-              beforeEach(async () => {
-                await userEvent.tab()
+                ).toHaveAttribute('aria-label', 'Hide day picker')
               })
 
-              it('focuses the picker container', () => {
-                expect(onCalendarFocus).toHaveBeenCalledTimes(1)
-                // NOTE: `react-day-picker` internals from here on down
+              describe('and the tab key is pressed again', () => {
+                beforeEach(() => {
+                  userEvent.tab()
+                })
+
+                it('focuses the picker container', () => {
+                  expect(onCalendarFocus).toHaveBeenCalledTimes(1)
+                  // NOTE: `react-day-picker` internals from here on down
+                })
               })
-            })
-          })
 
-          describe('and the return key is pressed', () => {
-            beforeEach(async () => {
-              const input = wrapper.getByTestId('datepicker-input')
+              describe('and clicks on a day', () => {
+                beforeEach(() => {
+                  userEvent.click(wrapper.getByText('31'))
+                })
 
-              await userEvent.type(input, '{enter}')
-            })
+                it('set the value of the component to this date', () => {
+                  expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+                    '31/05/2016'
+                  )
+                })
 
-            it('hides the day picker container', () => {
-              return waitFor(() => {
-                expect(wrapper.getByTestId('floating-box')).not.toBeVisible()
-              })
-            })
-          })
+                it('invokes the onChange callback', () => {
+                  expect(onChange).toHaveBeenLastCalledWith({
+                    startDate: new Date('2016-05-31T12:00:00.000Z'),
+                    startDateValidity: DATE_VALIDITY.VALID,
+                    endDate: new Date('2016-05-31T12:00:00.000Z'),
+                    endDateValidity: DATE_VALIDITY.VALID,
+                  })
+                })
 
-          describe('and clicks on a day', () => {
-            beforeEach(() => {
-              click(wrapper.getByText('31'))
-            })
-
-            it('set the value of the component to this date', () => {
-              expect(
-                wrapper.getByTestId('datepicker-input').getAttribute('value')
-              ).toBe('31/05/2016')
-            })
-
-            it('invokes the onChange callback', () => {
-              expect(onChange).toHaveBeenCalledTimes(2)
-              expect(onChange).toHaveBeenCalledWith({
-                startDate: new Date('2016-05-31T12:00:00.000Z'),
-                endDate: new Date('2016-05-31T12:00:00.000Z'),
-              })
-              expect(onChange).toHaveBeenCalledWith({
-                startDate: new Date('2016-05-01T12:00:00.000Z'),
-                endDate: new Date('2016-05-01T12:00:00.000Z'),
-              })
-            })
-
-            it('hides the day picker container', () => {
-              return waitFor(() => {
-                expect(wrapper.getByTestId('floating-box')).not.toBeVisible()
+                it('hides the day picker container', () => {
+                  return waitFor(() => {
+                    expect(
+                      wrapper.queryByTestId('floating-box')
+                    ).not.toBeInTheDocument()
+                  })
+                })
               })
             })
           })
         })
 
         describe('and the day/month use one digit format', () => {
-          beforeEach(async () => {
-            const input = wrapper.getByTestId('datepicker-input')
+          let input: HTMLElement
 
-            await userEvent.type(input, '{backspace}1/5/2016{enter}')
+          beforeEach(() => {
+            input = wrapper.getByTestId('datepicker-input')
+
+            userEvent.type(input, '{selectall}1/5/2016')
           })
 
-          it('set the value of the component to this date', () => {
-            expect(
-              wrapper.getByTestId('datepicker-input').getAttribute('value')
-            ).toBe('01/05/2016')
-          })
+          describe('and enter is pressed', () => {
+            beforeEach(() => {
+              userEvent.type(input, '{enter}')
+            })
 
-          it('invokes the `onChange` callback', () => {
-            expect(onChange).toHaveBeenCalledTimes(1)
-            expect(onChange).toHaveBeenCalledWith({
-              startDate: new Date('2016-05-01T12:00:00.000Z'),
-              endDate: new Date('2016-05-01T12:00:00.000Z'),
+            it('updates the value of the input with the formatted date', () => {
+              expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+                '01/05/2016'
+              )
             })
           })
 
-          it('updates the selected day', () => {
-            expect(wrapper.getByText('May 2016')).toBeInTheDocument()
-            expect(wrapper.getByText('1')).toBeInTheDocument()
+          describe('and tab is pressed', () => {
+            beforeEach(() => {
+              userEvent.tab()
+            })
+
+            it('updates the value of the input with the formatted date', () => {
+              expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+                '01/05/2016'
+              )
+            })
           })
 
-          it('hides the day picker container', () => {
-            expect(
-              wrapper.getByTestId('datepicker-input-button')
-            ).toHaveAttribute('aria-label', 'Show day picker')
+          describe('and the `input` loses `focus`', () => {
+            beforeEach(() => {
+              input.blur()
+            })
+
+            it('updates the value of the input with the formatted date', () => {
+              expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+                '01/05/2016'
+              )
+            })
           })
-        })
-      })
 
-      describe('and the escape key is pressed', () => {
-        beforeEach(async () => {
-          const input = wrapper.getByTestId('datepicker-input')
-          const newValue = '{backspace}20{esc}'
+          it('invokes the `onChange` callback', () => {
+            expect(onChange).toHaveBeenLastCalledWith({
+              startDate: new Date('2016-05-01T12:00:00.000Z'),
+              startDateValidity: DATE_VALIDITY.VALID,
+              endDate: new Date('2016-05-01T12:00:00.000Z'),
+              endDateValidity: DATE_VALIDITY.VALID,
+            })
+          })
 
-          await userEvent.type(input, newValue)
-        })
+          describe('and the open/close button is clicked', () => {
+            beforeEach(() => {
+              userEvent.click(wrapper.getByTestId('datepicker-input-button'))
+            })
 
-        it('revert the value of the component to the original date', () => {
-          expect(
-            wrapper.getByTestId('datepicker-input').getAttribute('value')
-          ).toBe('01/12/2019')
-        })
-
-        it('not invoke the `onChange` callback', () => {
-          expect(onChange).not.toHaveBeenCalled()
-        })
-
-        it('hides the day picker container', () => {
-          expect(
-            wrapper.getByTestId('datepicker-input-button')
-          ).toHaveAttribute('aria-label', 'Show day picker')
+            it('updates the selected day', () => {
+              expect(wrapper.getByText('May 2016')).toBeInTheDocument()
+              expect(wrapper.getByText('1')).toBeInTheDocument()
+            })
+          })
         })
       })
 
       describe('and a date is partially entered', () => {
+        beforeEach(() => {
+          const input = wrapper.getByTestId('datepicker-input')
+
+          userEvent.type(input, '{selectall}20')
+        })
         describe('and a day is selected', () => {
-          beforeEach(async () => {
-            const input = wrapper.getByTestId('datepicker-input')
-
-            await userEvent.type(input, '{backspace}20')
-
-            click(wrapper.getByText('21'))
+          beforeEach(() => {
+            userEvent.click(wrapper.getByTestId('datepicker-input-button'))
+            userEvent.click(wrapper.getByText('21'))
           })
 
           it('set the value of the component to the selected date', () => {
-            expect(
-              wrapper.getByTestId('datepicker-input').getAttribute('value')
-            ).toBe('21/12/2019')
+            expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+              '21/12/2019'
+            )
           })
 
           it('invokes the `onChange` callback', () => {
-            expect(onChange).toHaveBeenCalledTimes(1)
-            expect(onChange).toHaveBeenCalledWith({
+            expect(onChange).toHaveBeenLastCalledWith({
               startDate: new Date('2019-12-21T12:00:00.000Z'),
+              startDateValidity: DATE_VALIDITY.VALID,
               endDate: new Date('2019-12-21T12:00:00.000Z'),
+              endDateValidity: DATE_VALIDITY.VALID,
             })
           })
         })
@@ -437,178 +416,142 @@ describe('DatePicker', () => {
 
       describe('and the date value is invalid', () => {
         describe('when entering letters', () => {
-          describe('and using the enter key', () => {
-            beforeEach(async () => {
-              const input = wrapper.getByTestId('datepicker-input')
+          beforeEach(() => {
+            const input = wrapper.getByTestId('datepicker-input')
 
-              await userEvent.type(input, '{backspace}abcd{enter}')
-            })
-
-            it('not invoke the `onChange` callback', () => {
-              expect(onChange).not.toHaveBeenCalled()
-            })
-
-            it('should be in an error state', () => {
-              expect(
-                wrapper.getByTestId('datepicker-outer-wrapper')
-              ).toHaveStyleRule('border', ERROR_BORDER)
-            })
-
-            describe('and the escape key is pressed', () => {
-              beforeEach(async () => {
-                const input = wrapper.getByTestId('datepicker-input')
-
-                await userEvent.type(input, '{esc}')
-              })
-
-              it('not invoke the `onChange` callback', () => {
-                expect(onChange).not.toHaveBeenCalled()
-              })
-
-              it('should not be in an error state', () => {
-                expect(
-                  wrapper.getByTestId('datepicker-outer-wrapper')
-                ).not.toHaveStyleRule('border', ERROR_BORDER)
-              })
-            })
-
-            describe('and a valid date is entered', () => {
-              beforeEach(async () => {
-                const input = wrapper.getByTestId('datepicker-input')
-                const newValue =
-                  '{backspace}{backspace}{backspace}{backspace}01/03/2021{enter}'
-
-                await userEvent.type(input, newValue)
-              })
-
-              it('invokes the `onChange` callback', () => {
-                expect(onChange).toHaveBeenCalledTimes(1)
-                expect(onChange).toHaveBeenCalledWith({
-                  startDate: new Date('2021-03-01T12:00:00.000Z'),
-                  endDate: new Date('2021-03-01T12:00:00.000Z'),
-                })
-              })
-
-              it('should not be in an error state', () => {
-                expect(
-                  wrapper.getByTestId('datepicker-outer-wrapper')
-                ).not.toHaveStyleRule('border', ERROR_BORDER)
-              })
-            })
+            userEvent.type(input, '{selectall}abcd')
           })
 
-          describe('and tabbing', () => {
-            beforeEach(async () => {
-              const input = wrapper.getByTestId('datepicker-input')
+          it('calls the `onChange` callback with an invalid date', () => {
+            const lastCall = onChange.mock.calls[onChange.mock.calls.length - 1]
 
-              await userEvent.type(input, '{backspace}abcd')
+            expect(isValid(lastCall[0].startDate)).toBeFalsy()
+            expect(lastCall[0].startDateValidity).toBe(DATE_VALIDITY.INVALID)
+            expect(isValid(lastCall[0].endDate)).toBeFalsy()
+            expect(lastCall[0].endDateValidity).toBe(DATE_VALIDITY.INVALID)
+          })
 
+          describe('and the tab key is pressed', () => {
+            beforeEach(() => {
               userEvent.tab()
             })
 
-            it('not invoke the `onChange` callback', () => {
-              expect(onChange).not.toHaveBeenCalled()
-            })
-
-            it('should be in an error state', () => {
+            it('is in an error state', () => {
               expect(
                 wrapper.getByTestId('datepicker-outer-wrapper')
-              ).toHaveStyleRule('border', ERROR_BORDER)
+              ).toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
             })
 
-            it('should call the `onBlur` callback', () => {
+            it('calls the `onBlur` callback', () => {
               expect(onBlur).toHaveBeenCalledTimes(1)
             })
           })
 
-          describe('and the `input` loses `focus`', () => {
-            beforeEach(async () => {
+          describe('and a valid date is entered', () => {
+            beforeEach(() => {
               const input = wrapper.getByTestId('datepicker-input')
+              const newValue =
+                '{backspace}{backspace}{backspace}{backspace}01/03/2021'
 
-              await userEvent.type(input, '{backspace}abcd')
+              userEvent.type(input, newValue)
+            })
+
+            it('invokes the `onChange` callback', () => {
+              expect(onChange).toHaveBeenLastCalledWith({
+                startDate: new Date('2021-03-01T12:00:00.000Z'),
+                startDateValidity: DATE_VALIDITY.VALID,
+                endDate: new Date('2021-03-01T12:00:00.000Z'),
+                endDateValidity: DATE_VALIDITY.VALID,
+              })
+            })
+
+            it('should not be in an error state', () => {
+              expect(
+                wrapper.getByTestId('datepicker-outer-wrapper')
+              ).not.toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
+            })
+          })
+
+          describe('and the `input` loses `focus`', () => {
+            beforeEach(() => {
+              const input = wrapper.getByTestId('datepicker-input')
 
               input.blur()
             })
 
-            it('not invoke the `onChange` callback', () => {
-              expect(onChange).not.toHaveBeenCalled()
-            })
-
-            it('should be in an error state', () => {
+            it('is in an error state', () => {
               expect(
                 wrapper.getByTestId('datepicker-outer-wrapper')
-              ).toHaveStyleRule('border', ERROR_BORDER)
+              ).toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
             })
 
-            it('should call the `onBlur` callback', () => {
+            it('calls the `onBlur` callback', () => {
               expect(onBlur).toHaveBeenCalledTimes(1)
             })
           })
         })
 
         describe('when using a two digit year', () => {
-          beforeEach(async () => {
+          beforeEach(() => {
             const input = wrapper.getByTestId('datepicker-input')
-
-            await userEvent.type(input, '{backspace}12/12/20{enter}')
+            userEvent.type(input, '{selectall}12/12/20')
+            input.blur()
           })
 
-          it('not invoke the `onChange` callback', () => {
-            expect(onChange).not.toHaveBeenCalled()
+          it('calls the `onChange` callback with an invalid date', () => {
+            const lastCall = onChange.mock.calls[onChange.mock.calls.length - 1]
+
+            expect(isValid(lastCall[0].startDate)).toBeFalsy()
+            expect(lastCall[0].startDateValidity).toBe(DATE_VALIDITY.INVALID)
+            expect(isValid(lastCall[0].endDate)).toBeFalsy()
+            expect(lastCall[0].endDateValidity).toBe(DATE_VALIDITY.INVALID)
           })
 
-          it('should be in an error state', () => {
+          it('is in an error state', () => {
             expect(
               wrapper.getByTestId('datepicker-outer-wrapper')
-            ).toHaveStyleRule('border', ERROR_BORDER)
+            ).toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
           })
         })
 
         describe('when using a date that does not exist', () => {
-          beforeEach(async () => {
+          beforeEach(() => {
             const input = wrapper.getByTestId('datepicker-input')
-
-            await userEvent.type(input, '{backspace}15/15/20{enter}')
+            userEvent.type(input, '{selectall}15/15/20')
+            input.blur()
           })
 
-          it('not invoke the `onChange` callback', () => {
-            expect(onChange).not.toHaveBeenCalled()
+          it('calls the `onChange` callback with an invalid date', () => {
+            const lastCall = onChange.mock.calls[onChange.mock.calls.length - 1]
+
+            expect(isValid(lastCall[0].startDate)).toBeFalsy()
+            expect(lastCall[0].startDateValidity).toBe(DATE_VALIDITY.INVALID)
+            expect(isValid(lastCall[0].endDate)).toBeFalsy()
+            expect(lastCall[0].endDateValidity).toBe(DATE_VALIDITY.INVALID)
           })
 
-          it('should be in an error state', () => {
+          it('is in an error state', () => {
             expect(
               wrapper.getByTestId('datepicker-outer-wrapper')
-            ).toHaveStyleRule('border', ERROR_BORDER)
+            ).toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
           })
         })
       })
+    })
 
-      describe('and the `input` loses `focus`', () => {
-        beforeEach(async () => {
-          const input = wrapper.getByTestId('datepicker-input')
+    describe('when a new date is pasted', () => {
+      beforeEach(() => {
+        const input = wrapper.getByTestId('datepicker-input')
+        userEvent.clear(input)
+        userEvent.paste(input, '01/05/2021')
+      })
 
-          await userEvent.type(input, '{backspace}01/05/2016')
-
-          input.blur()
-        })
-
-        it('set the value of the component to this date', () => {
-          expect(
-            wrapper.getByTestId('datepicker-input').getAttribute('value')
-          ).toBe('01/05/2016')
-        })
-
-        it('invokes the `onChange` callback', () => {
-          expect(onChange).toHaveBeenCalledTimes(1)
-          expect(onChange).toHaveBeenCalledWith({
-            startDate: new Date('2016-05-01T12:00:00.000Z'),
-            endDate: new Date('2016-05-01T12:00:00.000Z'),
-          })
-        })
-
-        it('updates the selected day', () => {
-          expect(wrapper.getByText('May 2016')).toBeInTheDocument()
-          expect(wrapper.getByText('1')).toBeInTheDocument()
+      it('calls the `onChange` callback with the new date', () => {
+        expect(onChange).toHaveBeenLastCalledWith({
+          startDate: new Date('2021-05-01T12:00:00.000Z'),
+          startDateValidity: DATE_VALIDITY.VALID,
+          endDate: new Date('2021-05-01T12:00:00.000Z'),
+          endDateValidity: DATE_VALIDITY.VALID,
         })
       })
     })
@@ -622,7 +565,9 @@ describe('DatePicker', () => {
     })
 
     it('renders that label accordingly', () => {
-      expect(wrapper.getByTestId('datepicker-label').innerHTML).toBe(label)
+      expect(wrapper.getByTestId('datepicker-label').innerHTML).toBe(
+        `${label} (dd/mm/yyyy)`
+      )
     })
   })
 
@@ -645,12 +590,139 @@ describe('DatePicker', () => {
     })
   })
 
+  describe('when a single date picker is rendered and the picker is opened', () => {
+    beforeEach(() => {
+      wrapper = render(<DatePicker />)
+
+      userEvent.click(wrapper.getByTestId('datepicker-input-button'))
+    })
+
+    it('displays the current month', () => {
+      expect(wrapper.getByText('December 2019')).toBeInTheDocument()
+    })
+
+    it('focuses the previous month button', () => {
+      return waitFor(() =>
+        expect(wrapper.getByLabelText('Previous Month')).toHaveFocus()
+      )
+    })
+
+    describe('when the day picker is focused and Shift-Tab is pressed', () => {
+      let dayPicker: Element
+
+      beforeEach(() => {
+        dayPicker = wrapper.container.querySelectorAll('.DayPicker-wrapper')[0]
+
+        fireEvent.focus(dayPicker)
+
+        userEvent.tab({ shift: true })
+      })
+
+      it('traps the focus within the picker', () => {
+        expect(wrapper.getByText('1')).toHaveFocus()
+      })
+
+      describe('and Tab is then pressed', () => {
+        beforeEach(() => {
+          userEvent.tab()
+        })
+
+        it('still traps the focus within the picker', () => {
+          expect(dayPicker).toHaveFocus()
+        })
+      })
+    })
+
+    describe.each([
+      {
+        name: 'day picker container',
+        selector: () =>
+          wrapper.container.querySelectorAll('.DayPicker-wrapper')[0],
+      },
+      {
+        name: 'previous month button',
+        selector: () => wrapper.getByLabelText('Previous Month'),
+      },
+      {
+        name: 'day picker day',
+        selector: () => wrapper.getByText(10),
+      },
+    ])('when the escape key is pressed in $name', ({ selector }) => {
+      beforeEach(() => {
+        userEvent.type(selector(), '{esc}')
+      })
+
+      it('closes the day picker', () => {
+        return waitFor(() => {
+          expect(wrapper.queryByTestId('floating-box')).not.toBeInTheDocument()
+        })
+      })
+
+      it('focuses the open/close button', () => {
+        return waitFor(() =>
+          expect(wrapper.queryByTestId('datepicker-input-button')).toHaveFocus()
+        )
+      })
+    })
+
+    describe('when the next month button is clicked', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByLabelText('Next Month'))
+      })
+
+      it('displays the next month', () => {
+        expect(wrapper.getByText('January 2020')).toBeInTheDocument()
+      })
+
+      describe('when the first day is clicked', () => {
+        beforeEach(() => {
+          userEvent.click(wrapper.getByText('1'))
+        })
+
+        it('updates the input value', () => {
+          expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+            '01/01/2020'
+          )
+        })
+
+        it('closes the picker', () => {
+          return waitFor(() => {
+            expect(
+              wrapper.queryByTestId('floating-box')
+            ).not.toBeInTheDocument()
+          })
+        })
+
+        it('focuses the open/close button', () => {
+          return waitFor(() =>
+            expect(
+              wrapper.queryByTestId('datepicker-input-button')
+            ).toHaveFocus()
+          )
+        })
+
+        describe('when the picker is reopened', () => {
+          beforeEach(async () => {
+            await waitForElementToBeRemoved(
+              wrapper.queryByTestId('floating-box')
+            )
+
+            userEvent.click(wrapper.getByTestId('datepicker-input-button'))
+          })
+
+          it('opens the picker on the previously selected month', () => {
+            expect(wrapper.getByText('January 2020')).toBeInTheDocument()
+          })
+        })
+      })
+    })
+  })
+
   describe('when selecting a date range', () => {
     beforeEach(() => {
-      onChange = jest.fn()
       wrapper = render(
         <DatePicker
-          startDate={new Date(2019, 11, 10)}
+          initialStartDate={new Date(2019, 11, 10)}
           onChange={onChange}
           isRange
         />
@@ -670,96 +742,205 @@ describe('DatePicker', () => {
       )
     })
 
-    describe('when the end user focuses on the Input', () => {
+    describe('when the end user clicks on the Input', () => {
       beforeEach(() => {
-        wrapper.getByTestId('datepicker-input').focus()
+        userEvent.click(wrapper.getByTestId('datepicker-input'))
       })
 
-      describe('and clicks on a day in the first month', () => {
-        beforeEach(() => {
-          click(wrapper.getAllByText('31')[0])
-        })
-
-        it('set the value of the component to this date', () => {
-          expect(
-            wrapper.getByTestId('datepicker-input').getAttribute('value')
-          ).toBe('10/12/2019 - 31/12/2019')
-        })
-
-        it('invokes the onChange callback', () => {
-          expect(onChange).toHaveBeenCalledTimes(1)
-          expect(onChange).toHaveBeenCalledWith({
-            startDate: new Date('2019-12-10T00:00:00.000Z'),
-            endDate: new Date('2019-12-31T12:00:00.000Z'),
-          })
+      it('opens the day picker', () => {
+        return waitFor(() => {
+          expect(wrapper.getByTestId('floating-box')).toBeVisible()
         })
       })
 
-      describe('and clicks on days in the second month', () => {
+      describe('and hovers over a second date', () => {
         beforeEach(() => {
-          click(wrapper.getAllByText('20')[1])
+          userEvent.hover(wrapper.getByText('13'))
         })
 
-        it('set the value of the component to this date', () => {
+        it('shades the date range', () => {
           expect(
-            wrapper.getByTestId('datepicker-input').getAttribute('value')
-          ).toBe('10/12/2019 - 20/01/2020')
-        })
-
-        it('invokes the onChange callback', () => {
-          expect(onChange).toHaveBeenCalledTimes(1)
-          expect(onChange).toHaveBeenCalledWith({
-            startDate: new Date('2019-12-10T00:00:00.000Z'),
-            endDate: new Date('2020-01-20T12:00:00.000Z'),
-          })
-        })
-
-        describe('and clicks on another day', () => {
-          beforeEach(() => {
-            click(wrapper.getAllByText('3')[0])
-          })
-
-          it('reset the value of the component to this date', () => {
-            expect(
-              wrapper.getByTestId('datepicker-input').getAttribute('value')
-            ).toBe('03/12/2019')
-          })
-
-          it('invokes the onChange callback', () => {
-            expect(onChange).toHaveBeenCalledTimes(2)
-            expect(onChange).toHaveBeenCalledWith({
-              startDate: new Date('2019-12-03T12:00:00.000Z'),
-              endDate: undefined,
+            wrapper.getAllByText(/^10|11|12|13$/, {
+              selector: '.DayPicker-Day--selected',
             })
+          ).toHaveLength(4)
+        })
+
+        it("doesn't shade dates outside the range", () => {
+          expect(
+            wrapper.queryAllByText(/^(?!(10|11|12|13))\d\d$/, {
+              selector: '.DayPicker-Day--selected',
+            })
+          ).toHaveLength(0)
+        })
+
+        describe('and unhovers the second date', () => {
+          beforeEach(() => {
+            userEvent.unhover(wrapper.getByText('13'))
+          })
+
+          it('does not shade any days', () => {
+            expect(
+              wrapper.container.querySelectorAll('.DayPicker-Day--selected')
+            ).toHaveLength(0)
           })
         })
       })
 
-      describe('and clicks on a less than the first', () => {
+      describe('and focuses a second date', () => {
         beforeEach(() => {
-          click(wrapper.getAllByText('9')[0])
+          fireEvent.focus(wrapper.getByText('13'))
         })
 
-        it('reset the value of the component to this date', () => {
+        it('shades the date range', () => {
           expect(
-            wrapper.getByTestId('datepicker-input').getAttribute('value')
-          ).toBe('09/12/2019')
+            wrapper.getAllByText(/^10|11|12|13$/, {
+              selector: '.DayPicker-Day--selected',
+            })
+          ).toHaveLength(4)
+        })
+
+        it("doesn't shade dates outside the range", () => {
+          expect(
+            wrapper.queryAllByText(/^(?!(10|11|12|13))\d\d$/, {
+              selector: '.DayPicker-Day--selected',
+            })
+          ).toHaveLength(0)
+        })
+      })
+
+      describe('and clicks on a second date', () => {
+        beforeEach(() => {
+          userEvent.click(wrapper.getByText('20'))
+        })
+
+        it('closes the day picker', () => {
+          return waitFor(() => {
+            expect(
+              wrapper.queryByTestId('floating-box')
+            ).not.toBeInTheDocument()
+          })
+        })
+
+        it('focuses the open/close button', () => {
+          return waitFor(() =>
+            expect(
+              wrapper.queryByTestId('datepicker-input-button')
+            ).toHaveFocus()
+          )
+        })
+
+        it('set the value of the component to this date', () => {
+          expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+            '10/12/2019 - 20/12/2019'
+          )
+        })
+
+        it('invokes the onChange callback', () => {
+          expect(onChange).toHaveBeenCalledTimes(1)
+          expect(onChange).toHaveBeenCalledWith({
+            startDate: new Date('2019-12-10T00:00:00.000Z'),
+            startDateValidity: DATE_VALIDITY.VALID,
+            endDate: new Date('2019-12-20T12:00:00.000Z'),
+            endDateValidity: DATE_VALIDITY.VALID,
+          })
+        })
+      })
+
+      describe('and clicks on a date less than the first', () => {
+        beforeEach(() => {
+          userEvent.click(wrapper.getByText('9'))
+        })
+
+        it('closes the day picker', () => {
+          return waitFor(() => {
+            expect(
+              wrapper.queryByTestId('floating-box')
+            ).not.toBeInTheDocument()
+          })
+        })
+
+        it('updates the value to start from this date', () => {
+          expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+            '09/12/2019 - 10/12/2019'
+          )
         })
 
         it('invokes the onChange callback', () => {
           expect(onChange).toHaveBeenCalledTimes(1)
           expect(onChange).toHaveBeenCalledWith({
             startDate: new Date('2019-12-09T12:00:00.000Z'),
-            endDate: undefined,
+            startDateValidity: DATE_VALIDITY.VALID,
+            endDate: new Date('2019-12-10T00:00:00.000Z'),
+            endDateValidity: DATE_VALIDITY.VALID,
           })
         })
       })
     })
   })
 
-  describe('when the isOpen prop is provided', () => {
+  describe('when in range mode and two dates are already selected', () => {
+    let input: HTMLElement
+
     beforeEach(() => {
-      wrapper = render(<DatePicker isOpen />)
+      wrapper = render(
+        <DatePicker
+          initialStartDate={new Date(2019, 11, 10)}
+          initialEndDate={new Date(2019, 11, 15)}
+          isRange
+        />
+      )
+
+      input = wrapper.getByTestId('datepicker-input')
+    })
+
+    it('the input value displays the range selected', () => {
+      expect(input).toHaveValue('10/12/2019 - 15/12/2019')
+    })
+
+    describe('and the input is focused and blurred', () => {
+      beforeEach(() => {
+        input.focus()
+        input.blur()
+      })
+
+      it('the input value is unchanged', () => {
+        expect(input).toHaveValue('10/12/2019 - 15/12/2019')
+      })
+    })
+
+    // Note this shouldn't happen in reality as the input is read-only
+    // in range mode
+    describe('and return is pressed in the input', () => {
+      beforeEach(() => {
+        userEvent.type(input, '{enter}')
+      })
+
+      it('the input value is unchanged', () => {
+        expect(input).toHaveValue('10/12/2019 - 15/12/2019')
+      })
+    })
+
+    // Note this shouldn't happen in reality as the input is read-only
+    // in range mode
+    describe('and a date is typed in the input', () => {
+      beforeEach(() => {
+        fireEvent.change(input, {
+          target: {
+            value: '10/10/2021',
+          },
+        })
+      })
+
+      it('the input value is unchanged', () => {
+        expect(input).toHaveValue('10/12/2019 - 15/12/2019')
+      })
+    })
+  })
+
+  describe('when the initialIsOpen prop is provided', () => {
+    beforeEach(() => {
+      wrapper = render(<DatePicker initialIsOpen />)
     })
 
     it('displays the picker as open on initial render', () => {
@@ -802,13 +983,11 @@ describe('DatePicker', () => {
 
   describe('when the `disabledDays` prop is provided', () => {
     beforeEach(() => {
-      onChange = jest.fn()
-
       wrapper = render(
         <DatePicker
-          isOpen
+          initialIsOpen
           onChange={onChange}
-          startDate={new Date(2020, 3, 1)}
+          initialStartDate={new Date(2020, 3, 1)}
           disabledDays={[new Date(2020, 3, 12)]}
         />
       )
@@ -820,7 +999,9 @@ describe('DatePicker', () => {
 
     describe('and a disabled day is clicked', () => {
       beforeEach(() => {
-        click(wrapper.getByText('12'))
+        userEvent.click(wrapper.getByText('12'), undefined, {
+          skipPointerEventsCheck: true,
+        })
       })
 
       it('does not set the picker to that day', () => {
@@ -829,28 +1010,28 @@ describe('DatePicker', () => {
     })
 
     describe('and a disabled day is typed', () => {
-      beforeEach(async () => {
+      beforeEach(() => {
         const input = wrapper.getByTestId('datepicker-input')
 
-        await userEvent.type(input, '{backspace}12/04/2020{enter}')
+        userEvent.type(input, '{selectall}12/04/2020')
       })
 
-      it('does not set the picker to that day', () => {
-        expect(onChange).not.toHaveBeenCalled()
-      })
-
-      it('should be in an error state', () => {
-        expect(wrapper.getByTestId('datepicker-outer-wrapper')).toHaveStyleRule(
-          'border',
-          ERROR_BORDER
-        )
+      it('calls the `onChange` callback with the disabled date', () => {
+        expect(onChange).toHaveBeenLastCalledWith({
+          startDate: new Date('2020-04-12T12:00:00.000Z'),
+          startDateValidity: DATE_VALIDITY.DISABLED,
+          endDate: new Date('2020-04-12T12:00:00.000Z'),
+          endDateValidity: DATE_VALIDITY.DISABLED,
+        })
       })
     })
   })
 
   describe('when the `initialMonth` prop is provided and no `startDate`', () => {
     beforeEach(() => {
-      wrapper = render(<DatePicker isOpen initialMonth={new Date(2020, 1)} />)
+      wrapper = render(
+        <DatePicker initialIsOpen initialMonth={new Date(2020, 1)} />
+      )
     })
 
     it('displays the correct month initially', () => {
@@ -878,29 +1059,25 @@ describe('DatePicker', () => {
     })
   })
 
-  describe('when form validation classes are provided', () => {
-    describe('when the field is invalid', () => {
-      beforeEach(() => {
-        wrapper = render(<DatePicker className="is-invalid" />)
-      })
+  describe('when the is-invalid CSS class is set', () => {
+    beforeEach(() => {
+      wrapper = render(<DatePicker className="is-invalid" />)
+    })
 
-      it('should set the border on the outer wrapper', () => {
-        expect(wrapper.getByTestId('datepicker-outer-wrapper')).toHaveStyleRule(
-          'border',
-          ERROR_BORDER
-        )
-      })
+    it('should set the border on the outer wrapper', () => {
+      expect(wrapper.getByTestId('datepicker-outer-wrapper')).toHaveStyleRule(
+        'box-shadow',
+        ERROR_BOX_SHADOW
+      )
     })
   })
 
   describe('when `format` is specified', () => {
     beforeEach(() => {
-      onChange = jest.fn()
-
       wrapper = render(
         <DatePicker
           format="yyyy/MM/dd"
-          startDate={new Date(2018, 0, 11)}
+          initialStartDate={new Date(2018, 0, 11)}
           onChange={onChange}
         />
       )
@@ -918,25 +1095,26 @@ describe('DatePicker', () => {
     })
 
     describe('when a valid date is typed', () => {
-      beforeEach(async () => {
+      beforeEach(() => {
         const input = wrapper.getByTestId('datepicker-input')
 
-        userEvent.type(input, `2016/02/03{enter}`)
+        userEvent.type(input, `{selectall}2016/02/03`)
       })
 
       it('invokes the `onChange` callback', () => {
         const expectedDate = new Date('2016-02-03T12:00:00.000Z')
-        expect(onChange).toHaveBeenCalledTimes(1)
-        expect(onChange).toHaveBeenCalledWith({
+        expect(onChange).toHaveBeenLastCalledWith({
           startDate: expectedDate,
+          startDateValidity: DATE_VALIDITY.VALID,
           endDate: expectedDate,
+          endDateValidity: DATE_VALIDITY.VALID,
         })
       })
 
       it("isn't in an error state", () => {
         expect(
           wrapper.getByTestId('datepicker-outer-wrapper')
-        ).not.toHaveStyleRule('border', ERROR_BORDER)
+        ).not.toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
       })
     })
 
@@ -945,73 +1123,226 @@ describe('DatePicker', () => {
       (date) => {
         let input: HTMLElement
 
-        beforeEach(async () => {
+        beforeEach(() => {
           input = wrapper.getByTestId('datepicker-input')
 
           userEvent.type(input, date)
         })
 
-        describe.each([
-          ['enter', () => userEvent.type(input, '{enter}')],
-          ['tab', () => userEvent.tab()],
-        ])('and %s is pressed', (_, commitValue) => {
-          beforeEach(() => {
-            commitValue()
-          })
+        it('calls the `onChange` callback with an invalid date', () => {
+          const lastCall = onChange.mock.calls[onChange.mock.calls.length - 1]
 
-          it("doesn't call the `onChange` callback", () => {
-            expect(onChange).not.toHaveBeenCalled()
+          expect(isValid(lastCall[0].startDate)).toBeFalsy()
+          expect(lastCall[0].startDateValidity).toBe(DATE_VALIDITY.INVALID)
+          expect(isValid(lastCall[0].endDate)).toBeFalsy()
+          expect(lastCall[0].endDateValidity).toBe(DATE_VALIDITY.INVALID)
+        })
+
+        describe('and the input is blurred', () => {
+          beforeEach(() => {
+            input.blur()
           })
 
           it('is in an error state', () => {
             expect(
               wrapper.getByTestId('datepicker-outer-wrapper')
-            ).toHaveStyleRule('border', ERROR_BORDER)
+            ).toHaveStyleRule('box-shadow', ERROR_BOX_SHADOW)
           })
         })
       }
     )
   })
 
-  describe('when `startDate` and `endDate` are updated externally', () => {
+  describe('when a startDate and value is specified', () => {
+    // @ts-ignore
+    const jsxToRender = <DatePicker startDate={new Date()} value={null} />
+
+    it('does not throw an error', () => {
+      expect(() => render(jsxToRender)).not.toThrowError()
+    })
+  })
+
+  describe('when the state is held externally', () => {
+    let input: HTMLElement
+
+    const ExternalStateExample = () => {
+      const [value, setValue] = React.useState<Date | null>(null)
+
+      return (
+        <>
+          <DatePicker
+            startDate={value}
+            onChange={({ startDate: newStartDate }) => setValue(newStartDate)}
+          />
+          <div data-testid="value">{formatDate(value)}</div>
+          <button onClick={() => setValue(new Date())}>Today</button>
+          <button onClick={() => setValue(null)}>Clear date</button>
+        </>
+      )
+    }
+
     beforeEach(() => {
-      const initialProps = {
-        startDate: new Date(2021, 11, 1),
-        endDate: new Date(2021, 11, 2),
-      }
-      const update1Props = {
-        ...initialProps,
-        startDate: new Date(2022, 11, 1),
-      }
-      const update2Props = {
-        ...update1Props,
-        endDate: new Date(2022, 11, 2),
-      }
-
-      const DatePickerWithUpdate = () => {
-        const [props, updateProps] = useState(initialProps)
-
-        return (
-          <>
-            <Button onClick={() => updateProps(update1Props)}>Update 1</Button>
-            <Button onClick={() => updateProps(update2Props)}>Update 2</Button>
-            <DatePicker {...props} isRange />
-          </>
-        )
-      }
-
-      wrapper = render(<DatePickerWithUpdate />)
-
-      wrapper.getByText('Update 1').click()
-      wrapper.getByText('Update 2').click()
+      wrapper = render(<ExternalStateExample />)
+      input = wrapper.getByTestId('datepicker-input')
     })
 
-    it('set the value of the component to this date', () => {
-      return waitFor(() => {
-        expect(
-          wrapper.getByTestId('datepicker-input').getAttribute('value')
-        ).toBe('01/12/2022 - 02/12/2022')
+    describe('and a date is typed', () => {
+      beforeEach(() => {
+        userEvent.type(input, '06/05/2022')
       })
+
+      it('updates the input value', () => {
+        expect(input).toHaveValue('06/05/2022')
+      })
+
+      it('updates the external state', () => {
+        expect(wrapper.getByTestId('value')).toHaveTextContent('06/05/2022')
+      })
+    })
+
+    describe('and the external state is updated with a new date', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByText('Today'))
+      })
+
+      it('updates the input value', () => {
+        expect(input).toHaveValue('05/12/2019')
+      })
+
+      it('updates the external state', () => {
+        expect(wrapper.getByTestId('value')).toHaveTextContent('05/12/2019')
+      })
+
+      describe('and the external state is then updated with a null value', () => {
+        beforeEach(() => {
+          userEvent.click(wrapper.getByText('Clear date'))
+        })
+
+        it('updates the input value', () => {
+          expect(input).toHaveValue('')
+        })
+
+        it('updates the external state', () => {
+          expect(wrapper.getByTestId('value')).toBeEmptyDOMElement()
+        })
+      })
+    })
+  })
+
+  describe('when in range mode and the state is held externally', () => {
+    let input: HTMLElement
+
+    const RangeExternalStateExample = () => {
+      const [startDate, setStartDate] = useState<Date | null>(
+        new Date('2022-05-10')
+      )
+      const [endDate, setEndDate] = useState<Date | null>(
+        new Date('2022-05-20')
+      )
+
+      return (
+        <>
+          <DatePicker
+            startDate={startDate}
+            endDate={endDate}
+            onChange={({ startDate: newStartDate, endDate: newEndDate }) => {
+              setStartDate(newStartDate)
+              setEndDate(newEndDate)
+            }}
+            isRange
+          />
+          <div data-testid="start-date">{formatDate(startDate)}</div>
+          <div data-testid="end-date">{formatDate(endDate)}</div>
+          <button onClick={() => setStartDate(new Date('2022-05-01'))}>
+            Update start date
+          </button>
+          <button onClick={() => setEndDate(new Date('2022-06-01'))}>
+            Update end date
+          </button>
+          <button
+            onClick={() => {
+              setStartDate(null)
+              setEndDate(null)
+            }}
+          >
+            Clear dates
+          </button>
+        </>
+      )
+    }
+
+    beforeEach(() => {
+      wrapper = render(<RangeExternalStateExample />)
+      input = wrapper.getByTestId('datepicker-input')
+    })
+
+    describe('and the start date is updated externally', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByText('Update start date'))
+      })
+
+      it('updates the input value', () => {
+        expect(input).toHaveValue('01/05/2022 - 20/05/2022')
+      })
+
+      it('updates the external state', () => {
+        expect(wrapper.getByTestId('start-date')).toHaveTextContent(
+          '01/05/2022'
+        )
+        expect(wrapper.getByTestId('end-date')).toHaveTextContent('20/05/2022')
+      })
+    })
+
+    describe('and the end date is updated externally', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByText('Update end date'))
+      })
+
+      it('updates the input value', () => {
+        expect(input).toHaveValue('10/05/2022 - 01/06/2022')
+      })
+
+      it('updates the external state', () => {
+        expect(wrapper.getByTestId('start-date')).toHaveTextContent(
+          '10/05/2022'
+        )
+        expect(wrapper.getByTestId('end-date')).toHaveTextContent('01/06/2022')
+      })
+    })
+
+    describe('and the dates are cleared', () => {
+      beforeEach(() => {
+        userEvent.click(wrapper.getByText('Clear dates'))
+      })
+
+      it('updates the input value', () => {
+        expect(input).toHaveValue('')
+      })
+
+      it('updates the external state', () => {
+        expect(wrapper.getByTestId('start-date')).toBeEmptyDOMElement()
+        expect(wrapper.getByTestId('end-date')).toBeEmptyDOMElement()
+      })
+    })
+  })
+
+  describe('when initial and standard startDate and endDate props are provided', () => {
+    beforeEach(() => {
+      wrapper = render(
+        <DatePicker
+          startDate={new Date('2022-01-01')}
+          endDate={new Date('2022-02-28')}
+          initialStartDate={new Date('2019-05-01')}
+          initialEndDate={new Date('2019-06-31')}
+          isRange
+        />
+      )
+    })
+
+    it('prioritises startDate and endDate over initialStartDate and initialEndDate', () => {
+      expect(wrapper.getByTestId('datepicker-input')).toHaveValue(
+        '01/01/2022 - 28/02/2022'
+      )
     })
   })
 })
