+++ packages/react-component-library/src/components/DatePicker/index.ts
@@ -1,2 +1,1 @@
-export * from './constants'
 export * from './DatePicker'
