+++ packages/react-component-library/src/components/DatePicker/partials/StyledDatePickerInput.tsx
@@ -1,7 +1,6 @@
 import styled, { css } from 'styled-components'
-import { selectors } from '@defencedigital/design-tokens'
 
-const { spacing } = selectors
+import { StyledInput } from '../../TextInputE/partials/StyledInput'
 
 interface StyledDatePickerInputProps {
   $isDisabled?: boolean
@@ -12,12 +11,15 @@ export const StyledDatePickerInput = styled.div<StyledDatePickerInputProps>`
   flex-direction: column;
   position: relative;
   z-index: 0;
-  margin: ${spacing('6')} 0;
   padding: 0;
   border: 0;
   vertical-align: top;
   width: 100%;
   min-width: 16rem;
+
+  ${StyledInput}:not(:focus)::placeholder {
+    color: transparent;
+  }
 
   ${({ $isDisabled }) =>
     $isDisabled &&
