+++ packages/react-component-library/src/components/DatePicker/partials/StyledFloatingBox.tsx
