+++ packages/react-component-library/src/components/DatePicker/partials/StyledSeparator.tsx
@@ -1,11 +1,10 @@
 import styled from 'styled-components'
 import { selectors } from '@defencedigital/design-tokens'
 
-const { color, spacing } = selectors
+const { color } = selectors
 
 export const StyledSeparator = styled.div`
   width: 1px;
-  height: 100%;
-  margin-right: ${spacing('8')};
-  background: ${color('neutral', '100')};
+  align-self: stretch;
+  background: ${color('neutral', '200')};
 `
