+++ packages/react-component-library/src/components/DatePicker/types.ts
@@ -1,4 +1,40 @@
-export interface StateObject {
-  from?: Date
-  to?: Date
+export interface DatePickerState {
+  startDate: Date | null
+  endDate: Date | null
+  inputValue: string
+  datePickerFormat: string
+  hasError: boolean
 }
+
+export const DATEPICKER_E_ACTION = {
+  RESET: 'reset',
+  UPDATE: 'update',
+  REFRESH_HAS_ERROR: 'refreshHasError',
+  REFRESH_INPUT_VALUE: 'refreshInputValue',
+} as const
+
+interface ResetAction {
+  type: typeof DATEPICKER_E_ACTION.RESET
+  data: Omit<DatePickerState, 'hasError' | 'inputValue'>
+}
+
+interface UpdateAction {
+  type: typeof DATEPICKER_E_ACTION.UPDATE
+  data: Partial<DatePickerState>
+}
+
+interface RefreshHasErrorAction {
+  type: typeof DATEPICKER_E_ACTION.REFRESH_HAS_ERROR
+  data?: never
+}
+
+interface RefreshInputValueAction {
+  type: typeof DATEPICKER_E_ACTION.REFRESH_INPUT_VALUE
+  data?: never
+}
+
+export type DatePickerAction =
+  | ResetAction
+  | UpdateAction
+  | RefreshHasErrorAction
+  | RefreshInputValueAction
