+++ packages/react-component-library/src/components/DatePicker/useFocusTrapOptions.ts
