+++ packages/react-component-library/src/components/DatePicker/useRangeHoverOrFocusDate.ts
