+++ packages/react-component-library/src/components/DatePicker/utils/areDatesEqual.test.ts
