+++ packages/react-component-library/src/components/DatePicker/utils/areDatesEqual.ts
