+++ packages/react-component-library/src/components/DatePicker/utils/formatDatesForInput.ts
