--- packages/react-component-library/src/components/DatePickerE/index.ts
@@ -1,1 +1,0 @@
-export * from './DatePickerE'
