--- packages/react-component-library/src/components/DatePickerE/partials/StyledInputWrapper.tsx
@@ -1,18 +1,0 @@
-import styled, { css } from 'styled-components'
-
-interface StyledInputWrapperProps {
-  $isRange: boolean
-}
-
-export const StyledInputWrapper = styled.div<StyledInputWrapperProps>`
-  position: relative;
-  flex-grow: 1;
-
-  ${({ $isRange }) =>
-    $isRange &&
-    css`
-      input {
-        cursor: default;
-      }
-    `}
-`
