--- packages/react-component-library/src/components/DatePickerE/partials/StyledOuterWrapper.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-import { StyledOuterWrapper as StyledOuterWrapperBase } from '../../../styled-components'
-
-export const StyledOuterWrapper = styled(StyledOuterWrapperBase)`
-  display: inline-flex;
-  flex-direction: row;
-`
