+++ packages/react-component-library/src/forms/native/native.stories.tsx
@@ -7,7 +7,7 @@ import { TextAreaE } from '../../components/TextAreaE'
 import { Radio } from '../../components/Radio'
 import { Autocomplete, AutocompleteOption } from '../../components/Autocomplete'
 import { Checkbox } from '../../components/Checkbox'
-import { DatePickerE } from '../../components/DatePickerE'
+import { DatePicker } from '../../components/DatePicker'
 import { NumberInputE } from '../../components/NumberInputE'
 import { SelectE, SelectEOption } from '../../components/SelectE'
 import { SwitchE, SwitchEOption } from '../../components/SwitchE'
@@ -194,7 +194,7 @@ export const Example: React.FC<unknown> = () => {
           hintText="Example hint text."
           errors={[{ error: formErrors.exampleDatePicker }]}
         >
-          <DatePickerE
+          <DatePicker
             disabledDays={{ before: MINIMUM_DATE }}
             onChange={({ startDate }) => {
               handleChange({
