+++ packages/react-component-library/src/index.ts
@@ -20,7 +20,6 @@ export * from './components/CardFrame'
 export * from './components/Checkbox'
 export * from './components/ContextMenu'
 export * from './components/DatePicker'
-export * from './components/DatePickerE'
 export * from './components/DescriptionList'
 export * from './components/Dialog'
 export * from './components/DismissibleBanner'
